+++ src/services/client.js
@@ -1,10 +1,10 @@
 import axios from 'axios';
 
-// eslint-disable-next-line import/prefer-default-export
-export const getPartner = () => {
-  axios.get(`https://fasttrack-octobre-back.herokuapp.com/api/partner/${partnerId}`)
-    .then((res) => {
-      const { data } = res;
-      return data;
-    });
+export const dataSender = (formData) => {
+  axios({
+    method: 'post',
+    url: 'https://fasttrack-octobre-back.herokuapp.com/api/partner',
+    data: formData,
+    headers: { 'Content-Type': 'multipart/form-data' },
+  });
 };
