+++ src/services/theme.js
@@ -1,5 +1,5 @@
 export const theme = {
   colors: {
-    primary: 'hotpink'
-  }
-}
+    primary: 'hotpink',
+  },
+};
