+++ src/store/index.js
@@ -1,23 +1,6 @@
-// == Import : npm
-import { createStore, compose, applyMiddleware } from 'redux';
+import { combineReducers } from 'redux';
+import partnerReducer from './reducers/partner';
 
-// == Import : local
-import reducer from './reducer';
-import logMiddleware from './logMiddleware';
-
-// == Store
-const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
-
-const enhancers = composeEnhancers(
-  applyMiddleware(
-    logMiddleware,
-  ),
-);
-
-const store = createStore(
-  reducer,
-  enhancers,
-);
-
-// == Export
-export default store;
+export default combineReducers({
+  partnerReducer,
+});
