--- src/store/reducer.js
@@ -1,30 +1,0 @@
-// == Initial State
-const initialState = {
-  message: 'Hello',
-};
-
-// == Types
-const DO_SOMETHING = 'DO_SOMETHING';
-
-// == Reducer
-const reducer = (state = initialState, action = {}) => {
-  switch (action.type) {
-    case DO_SOMETHING:
-      return {
-        ...state,
-        message: action.message,
-      };
-
-    default:
-      return state;
-  }
-};
-
-// == Action Creators
-export const doSomething = (message) => ({
-  type: DO_SOMETHING,
-  message,
-});
-
-// == Export
-export default reducer;
